+++ src/mainboard/google/nyan/boardid.c
@@ -17,7 +17,7 @@
 #include <console/console.h>
 #include <gpio.h>
 
-uint8_t board_id(void)
+uint32_t board_id(void)
 {
 	static int id = -1;
